gh-235 Rename Sqlite to EntityFramework
