+++ src/Database/EntityFramework/InformaticsGatewayContext.cs
@@ -18,9 +18,9 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.Extensions.Logging;
 using Monai.Deploy.InformaticsGateway.Api;
-using Monai.Deploy.InformaticsGateway.Database.Sqlite.Configurations;
+using Monai.Deploy.InformaticsGateway.Database.EntityFramework.Configurations;
 
-namespace Monai.Deploy.InformaticsGateway.Database.Sqlite
+namespace Monai.Deploy.InformaticsGateway.Database.EntityFramework
 {
     public class InformaticsGatewayContext : DbContext
     {
