+++ src/InformaticsGateway/Program.cs
@@ -29,8 +29,6 @@ using Microsoft.Extensions.Options;
 using Monai.Deploy.InformaticsGateway.Common;
 using Monai.Deploy.InformaticsGateway.Configuration;
 using Monai.Deploy.InformaticsGateway.Database;
-using Monai.Deploy.InformaticsGateway.Database.Api;
-using Monai.Deploy.InformaticsGateway.Database.Sqlite;
 using Monai.Deploy.InformaticsGateway.Repositories;
 using Monai.Deploy.InformaticsGateway.Services.Common;
 using Monai.Deploy.InformaticsGateway.Services.Connectors;
@@ -112,7 +110,6 @@ namespace Monai.Deploy.InformaticsGateway
                     services.AddTransient<IStreamsWriter, StreamsWriter>();
                     services.AddTransient<IApplicationEntityHandler, ApplicationEntityHandler>();
 
-                    services.AddScoped(typeof(IInformaticsGatewayRepository<>), typeof(InformaticsGatewayRepository<>));
                     services.AddScoped<IStorageMetadataWrapperRepository, StorageMetadataWrapperRepository>();
                     services.AddScoped<IInferenceRequestRepository, InferenceRequestRepository>();
                     services.AddScoped<IPayloadMoveActionHandler, PayloadMoveActionHandler>();
