+++ src/InformaticsGateway/Test/Repositories/InformaticsGatewayRepositoryTest.cs
@@ -21,8 +21,7 @@ using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Monai.Deploy.InformaticsGateway.Api;
-using Monai.Deploy.InformaticsGateway.Database.Sqlite;
-using Monai.Deploy.InformaticsGateway.Repositories;
+using Monai.Deploy.InformaticsGateway.Database.EntityFramework;
 using Moq;
 using xRetry;
 using Xunit;
