+++ tests/Integration.Test/Hooks/SqlHooks.cs
@@ -16,7 +16,7 @@
 
 using Microsoft.EntityFrameworkCore;
 using Monai.Deploy.InformaticsGateway.Api.Rest;
-using Monai.Deploy.InformaticsGateway.Database.Sqlite;
+using Monai.Deploy.InformaticsGateway.Database.EntityFramework;
 using Monai.Deploy.InformaticsGateway.Integration.Test.Drivers;
 using Monai.Deploy.InformaticsGateway.Integration.Test.StepDefinitions;
 using TechTalk.SpecFlow.Infrastructure;
